Projects first commit
This commit is the first commit for this project, this commit contains the "Replace" library function for the C++ language which allows the replacement of strings.

Commit also provides the initial file structure for the project repository, aswell as example C/C++ programs to demonstrate the usage of the library.