Within this commit, the trimming library function has been implemented within the stringoperation library. This library function allows C++ string objects to be trimmed, trimming is the removal of leading and trailing whitespaces within a string object.
The function requires two function parameters, the first parameter is the
string object which is going to be trimmed and the second parameter is the
trimming operator value, which can be either left trim (L), right trim (R)
and trim all (A).

All specific information on how to use this library function will be provided
in the project documentation.